chore: move database errors
