+++ go.mod
@@ -3,6 +3,6 @@ module github.com/bnb-chain/bas-smt
 go 1.14
 
 require (
-	github.com/ethereum/go-ethereum v1.10.17 // indirect
-	github.com/syndtr/goleveldb v1.0.1-0.20210819022825-2ae1ddf74ef7 // indirect
+	github.com/ethereum/go-ethereum v1.10.17
+	github.com/syndtr/goleveldb v1.0.1-0.20210819022825-2ae1ddf74ef7
 )
